+++ src/CompoundCurve.php
@@ -15,8 +15,10 @@ use Override;
 
 /**
  * A CompoundCurve is a collection of zero or more continuous CircularString or LineString instances.
+ *
+ * @template-implements \IteratorAggregate<int<0, max>, LineString|CircularString>
  */
-class CompoundCurve extends Curve
+class CompoundCurve extends Curve implements \Countable, \IteratorAggregate
 {
     /**
      * The Curves that compose this CompoundCurve.
@@ -187,8 +189,6 @@ class CompoundCurve extends Curve
 
     /**
      * Returns the number of curves in this CompoundCurve.
-     *
-     * Required by interface Countable.
      */
     #[Override]
     public function count() : int
@@ -199,9 +199,7 @@ class CompoundCurve extends Curve
     /**
      * Returns an iterator for the curves in this CompoundCurve.
      *
-     * Required by interface IteratorAggregate.
-     *
-     * @psalm-return ArrayIterator<int<0, max>, LineString|CircularString>
+     * @return ArrayIterator<int<0, max>, LineString|CircularString>
      */
     #[Override]
     public function getIterator() : ArrayIterator