+++ src/LineString.php
@@ -17,8 +17,10 @@ use Override;
  * A LineString is a Curve with linear interpolation between Points.
  *
  * Each consecutive pair of Points defines a line segment.
+ *
+ * @template-implements \IteratorAggregate<int<0, max>, Point>
  */
-class LineString extends Curve
+class LineString extends Curve implements \Countable, \IteratorAggregate
 {
     /**
      * The Points that compose this LineString.
@@ -26,9 +28,7 @@ class LineString extends Curve
      * An empty LineString contains no points.
      * A non-empty LineString contains a minimum of 2 points.
      *
-     * @psalm-var list<Point>
-     *
-     * @var Point[]
+     * @var list<Point>
      */
     protected array $points = [];
 
@@ -162,9 +162,7 @@ class LineString extends Curve
     /**
      * Returns the points that compose this LineString.
      *
-     * @psalm-return list<Point>
-     *
-     * @return Point[]
+     * @return list<Point>
      */
     public function points() : array
     {
@@ -221,8 +219,6 @@ class LineString extends Curve
 
     /**
      * Returns the number of points in this LineString.
-     *
-     * Required by interface Countable.
      */
     #[Override]
     public function count() : int
@@ -233,9 +229,7 @@ class LineString extends Curve
     /**
      * Returns an iterator for the points in this LineString.
      *
-     * Required by interface IteratorAggregate.
-     *
-     * @psalm-return ArrayIterator<int, Point>
+     * @return ArrayIterator<int<0, max>, Point>
      */
     #[Override]
     public function getIterator() : ArrayIterator