+++ src/Proxy/CurveProxy.php
@@ -298,22 +298,4 @@ class CurveProxy extends Curve implements ProxyInterface
         return $this->proxyGeometry->isIdenticalTo($that);
     }
 
-    public function count()
-    {
-        if ($this->proxyGeometry === null) {
-            $this->load();
-        }
-
-        return $this->proxyGeometry->count();
-    }
-
-    public function getIterator() : \Traversable
-    {
-        if ($this->proxyGeometry === null) {
-            $this->load();
-        }
-
-        return $this->proxyGeometry->getIterator();
-    }
-
 }