+++ src/Proxy/GeometryProxy.php
@@ -288,22 +288,4 @@ class GeometryProxy extends Geometry implements ProxyInterface
         return $this->proxyGeometry->isIdenticalTo($that);
     }
 
-    public function count()
-    {
-        if ($this->proxyGeometry === null) {
-            $this->load();
-        }
-
-        return $this->proxyGeometry->count();
-    }
-
-    public function getIterator() : \Traversable
-    {
-        if ($this->proxyGeometry === null) {
-            $this->load();
-        }
-
-        return $this->proxyGeometry->getIterator();
-    }
-
 }