+++ src/Triangle.php
@@ -14,20 +14,19 @@ use Override;
  */
 class Triangle extends Polygon
 {
-    public function __construct(CoordinateSystem $cs, LineString ...$rings)
+    #[Override]
+    protected function validate(): void
     {
-        parent::__construct($cs, ...$rings);
-
         if ($this->isEmpty) {
             return;
         }
 
         if ($this->exteriorRing()->numPoints() !== 4) {
-            throw new InvalidGeometryException('A triangle must have exactly 4 (3 + 1) points.');
+            throw new InvalidGeometryException('A Triangle must have exactly 4 (3 + first again) points.');
         }
 
         if ($this->numInteriorRings() !== 0) {
-            throw new InvalidGeometryException('A triangle must not have interior rings.');
+            throw new InvalidGeometryException('A Triangle must not have interior rings.');
         }
     }