+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Interface `GeometryEngine` has 2 new methods: `lineInterpolatePoint()`, `lineInterpolatePoints()`
 - `BoundingBox`'s constructor is now private, use `BoundingBox::new()` to create a new instance
 - `CompoundCurve` can now only contain `LineString` and `CircularString` instances; nested `CompoundCurve` instances are forbidden
+- `Geometry` no longer implements `Countable` and `IteratorAggregate`; subclasses implement these interfaces as needed, with proper types
 - The following classes are now `final`:
   - `BoundingBox`
   - `CoordinateSystem`